src/requirements.txt
#eveonline login 
certifi==2018.8.24
chardet==3.0.4
ecdsa==0.13.3
future==0.16.0
idna==2.7
pyasn1==0.4.4
python-jose==3.0.1
requests==2.27.1
rsa==3.4.2
six==1.11.0
urllib3==1.24.2
jose
#Webapp
Flask==2.0.2
flask-login
flask-cors==3.0.10
#requests==2.27.1
requests-toolbelt==0.9.1
google-cloud-datastore==2.4.0
firebase-admin
google-auth